+++ kikuchipy/io/plugins/tests/test_h5ebsd.py
@@ -288,9 +288,10 @@ class Testh5ebsd:
 
     def test_load_readonly(self):
         s = load(KIKUCHIPY_FILE, lazy=True)
+        keys = ["array-original", "original-array"]
         k = next(
             filter(
-                lambda x: isinstance(x, str) and x.startswith("array-original"),
+                lambda x: isinstance(x, str) and any([x.startswith(j) for j in keys]),
                 s.data.dask.keys(),
             )
         )