+++ kikuchipy/projections/gnomonic_projection.py
@@ -97,6 +97,6 @@ class GnomonicProjection(SphericalProjection):
         >>> xyz = GnomonicProjection.xy2vector(xy_g)
         """
         x, y = xy[..., 0], xy[..., 1]
-        polar = np.arctan(np.sqrt(x ** 2 + y ** 2))
+        polar = np.arctan(np.sqrt(x**2 + y**2))
         azimuth = np.arctan2(y, x)
         return Vector3d.from_polar(polar=polar, azimuth=azimuth)